Move schemas.ts and SchemaDetails to /internal
