+++ packages/components/src/internal/components/domainproperties/DomainForm.tsx
@@ -29,7 +29,7 @@ import { FIELD_EDITOR_TOPIC, helpLinkNode } from '../../util/helpLinks';
 
 import { blurActiveElement } from '../../util/utils';
 
-import { EXPAND_TRANSITION, EXPAND_TRANSITION_FAST, PHILEVEL_NOT_PHI, SEVERITY_LEVEL_ERROR } from './constants';
+import { DEFAULT_DOMAIN_FORM_DISPLAY_OPTIONS, EXPAND_TRANSITION, EXPAND_TRANSITION_FAST, PHILEVEL_NOT_PHI, SEVERITY_LEVEL_ERROR } from './constants';
 import { LookupProvider } from './Lookup/Context';
 import {
     addDomainField,
@@ -52,7 +52,6 @@ import {
     DomainField,
     DomainFieldError,
     DomainPanelStatus,
-    DEFAULT_DOMAIN_FORM_DISPLAY_OPTIONS,
     HeaderRenderer,
     IAppDomainHeader,
     IDomainField,
