+++ packages/components/src/internal/components/user/UserProfile.spec.tsx
@@ -4,11 +4,8 @@ import { mount } from 'enzyme';
 import { Button } from 'react-bootstrap';
 
 import { TEST_USER_READER } from '../../../test/data/users';
-import { FileInput } from '../forms/input/FileInput';
-import { TextInput } from '../forms/input/TextInput';
-import { getQueryDetails } from '../../query/api';
+import { getQueryDetails, FileInput, SCHEMAS, TextInput } from '../../..';
 import { initUnitTestMocks } from '../../testHelpers';
-import { SCHEMAS } from '../base/models/schemas';
 
 import { UserProfile } from './UserProfile';
