+++ packages/components/src/internal/util/AppURLResolver.ts
@@ -16,12 +16,7 @@
 import { List, Map } from 'immutable';
 import { Filter } from '@labkey/api';
 
-import { AssayProtocolModel } from '../components/domainproperties/assay/models';
-import { fetchProtocol } from '../components/domainproperties/assay/actions';
-
-import { getQueryDetails, selectRows } from '../query/api';
-import { AppURL, spliceURL } from '../url/AppURL';
-import { SCHEMAS } from '../components/base/models/schemas';
+import { AppURL, AssayProtocolModel, fetchProtocol, getQueryDetails, SCHEMAS, selectRows, spliceURL } from '../..';
 
 export interface AppRouteResolver {
     matches: (route: string) => boolean;
